Moving solution explorer types to apis-products
