+++ pkg/apis-products/tutorial-web-app-operator/v1alpha1/register.go
