+++ pkg/products/solutionexplorer/reconciler.go
@@ -18,7 +18,7 @@ import (
 	"github.com/integr8ly/integreatly-operator/pkg/products/monitoring"
 
 	monitoringv1alpha1 "github.com/integr8ly/application-monitoring-operator/pkg/apis/applicationmonitoring/v1alpha1"
-	webapp "github.com/integr8ly/integreatly-operator/pkg/apis/integreatly/tutorial-web-app-operator/pkg/apis/v1alpha1"
+	solutionExplorerv1alpha1 "github.com/integr8ly/integreatly-operator/pkg/apis-products/tutorial-web-app-operator/v1alpha1"
 	integreatlyv1alpha1 "github.com/integr8ly/integreatly-operator/pkg/apis/integreatly/v1alpha1"
 	"github.com/integr8ly/integreatly-operator/pkg/config"
 	"github.com/integr8ly/integreatly-operator/pkg/resources"
@@ -349,7 +349,7 @@ func (r *Reconciler) ReconcileCustomResource(ctx context.Context, installation *
 	if err != nil {
 		return integreatlyv1alpha1.PhaseFailed, err
 	}
-	seCR := &webapp.WebApp{
+	seCR := &solutionExplorerv1alpha1.WebApp{
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace: r.Config.GetNamespace(),
 			Name:      defaultName,
